requirements.txt
numpy
sounddevice
soundfile
openai-whisper
